+++ source/compress_integer_simple_9_packed.cpp
@@ -166,8 +166,8 @@ namespace JASS
 		while (pos >= 0)
 			{
 			size_t remaining = (pos + 28 < static_cast<int64_t>(source_integers)) ? 28 : source_integers - pos;
-			size_t last_bitmask = 0x0000;
-			size_t bitmask = 0xFFFF;
+			uint16_t last_bitmask = 0x0000;
+			uint16_t bitmask = 0xFFFF;
 			/*
 				constrain last_bitmask to contain only bits for masks we can pack with
 			*/
@@ -186,7 +186,7 @@ namespace JASS
 			/*
 				iterate through the bitmask bit-wise and try the optimization.
 			*/
-			for (size_t offset = 0; offset < 9; offset++)
+			for (uint8_t offset = 0; offset < 9; offset++)
 				if ((1 << offset) & last_bitmask)
 					{
 					size_t num_to_pack = ints_packed_table[offset];
@@ -227,7 +227,7 @@ namespace JASS
 			if (into + 1 > end)
 				return 0;
 				
-			size_t mask_type = masks[pos];
+			uint32_t mask_type = masks[pos];
 			size_t num_to_pack = (pos + ints_packed_table[mask_type] > source_integers) ? source_integers - pos : ints_packed_table[mask_type];
 			/*
 				pack the word
